Use the current rvm instead of default
Do not need to wrap this in bundle exec (it is not a gem script/binary). Use `current` instead of `default` with `do ruby` so that we get the ruby version associated with the project trying to make the commit instead of the default ruby, which might be a mismatch and cause the pre-commit script to blow up.